requirements.txt
# requirements.txt
# This file lists the Python packages required to run the
# data preparation, model training pipeline, and the EDA notebook.

# Core Pipeline Requirements
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.1.4
pyyaml==6.0.1
joblib==1.4.2

# EDA Notebook Requirements
# These packages are typically needed for running Jupyter notebooks and plotting.
ipykernel==6.29.5 # Essential for running Jupyter notebooks
matplotlib==3.8.4
seaborn==0.13.2
jupyter_client==8.6.3 # Required by ipykernel
jupyter_core==5.7.2  # Required by ipykernel
Pygments==2.19.1     # Syntax highlighting in notebooks
prompt_toolkit==3.0.50 # Used by IPython/Jupyter console
traitlets==5.14.3    # Dependency for Jupyter/IPython
wcwidth==0.2.13      # Dependency for prompt_toolkit
decorator==5.1.1     # Common dependency for interactive environments
nest-asyncio==1.6.0  # For running async code in notebooks
comm==0.2.2          # Dependency for ipykernel
parso==0.8.4         # Python parser used by Jedi
jedi==0.19.2         # Autocompletion for IPython/Jupyter
pickleshare==0.7.5   # Simple datastore used by IPython

# Dependencies often installed transitively, but explicitly listed for clarity if issues arise
# python-dateutil==2.9.0.post0 # Dependency of pandas
# pytz==2024.2                 # Dependency of pandas
# tzdata==2025.1               # Dependency of pandas
# scipy==1.15.1                # Dependency of scikit-learn
# threadpoolctl==3.5.0         # Dependency of scikit-learn

# Python Version Requirement:
# This project is developed and tested with Python 3.11.
# It is recommended to use Python 3.11 for compatibility.